+++ client/src/app/base/base.module.ts
@@ -1,0 +1,18 @@
+/**
+ * Module with shared componets for all application.
+ * @module base.module
+ */
+
+// Dependencies
+import * as angular from 'angular'
+
+// Feature modules
+
+/**
+ * Module name
+ */
+export const baseModule = 'base'
+
+// Define the module
+angular
+  .module(baseModule, [])