+++ client/src/app/base/index.ts
@@ -1,4 +1,2 @@
-/** Group base components */
-// Put in this folder components that are shared in all aplication or aren'that
-// a root component
-// export { header } from './header/index'
+/** base module */
+export { baseModule } from './base.module'