+++ client/src/app/shared/directives/directives.module.ts
@@ -1,0 +1,20 @@
+/**
+ * Group shared directives for all app.
+ * @module directives
+ */
+
+// Dependencies
+import * as angular from 'angular'
+
+// Directives
+// import { closeDdDirective, closeDdDirectiveName } from './close-dd.directive'
+
+/**
+ * Module name
+ */
+export const directivesModule = 'directives'
+
+// Define module
+angular
+  .module(directivesModule, [])
+  // .directive(closeDdDirectiveName, closeDdDirective)