+++ client/src/app/shared/interceptor/interceptor.module.ts
@@ -7,14 +7,15 @@
 import * as angular from 'angular'
 // import { errorInterceptorFactory } from './error-interceptor.factory'
 
-/** Module name */
-export const interceptor = 'interceptor'
+/**
+ * Module name
+ */
+export const interceptorModule = 'interceptor'
 
 // Define module
 angular
-  .module(interceptor, [])
+  .module(interceptorModule, [])
   .config(['$httpProvider', $httpProvider => {
     // Attach error interceptor service to the http response
-
     // $httpProvider.interceptors.push(errorInterceptorFactory)
   }])